refactor: extract mapFinishReasonToStopReason to shared utility
Address CodeRabbit review nits:
- Extract duplicated mapFinishReasonToStopReason into finishReasonMapping.ts
- Fix misleading test name for end_turn stopReason test case
