+++ packages/core/src/providers/openai/OpenAINonStreamHandler.ts
@@ -33,24 +33,7 @@ import {
   sanitizeToolArgumentsString,
   extractKimiToolCallsFromText,
 } from './OpenAIResponseParser.js';
-
-/**
- * Map OpenAI finish_reason to the stopReason format expected by MessageConverter.
- * OpenAI values: stop, length, tool_calls, content_filter
- * MessageConverter expects: end_turn, max_tokens, stop_sequence, tool_use, etc.
- */
-function mapFinishReasonToStopReason(
-  finishReason: string | null | undefined,
-): string | undefined {
-  if (!finishReason) return undefined;
-  const mapping: Record<string, string> = {
-    stop: 'end_turn',
-    length: 'max_tokens',
-    tool_calls: 'tool_use',
-    content_filter: 'end_turn',
-  };
-  return mapping[finishReason] ?? finishReason;
-}
+import { mapFinishReasonToStopReason } from './finishReasonMapping.js';
 
 export interface NonStreamHandlerDeps {
   toolCallPipeline: ToolCallPipeline;
